+++ code/widgets/org.eclipse.scout.widgets.old.client/src/main/java/org/eclipse/scout/widgets/old/client/ui/forms/BrowserFieldForm.java
@@ -237,7 +237,7 @@ public class BrowserFieldForm extends AbstractForm implements IPageForm {
           }
 
           @Override
-          protected int getConfiguredLabelPosition() {
+          protected byte getConfiguredLabelPosition() {
             return LABEL_POSITION_ON_FIELD;
           }
         }
@@ -251,7 +251,7 @@ public class BrowserFieldForm extends AbstractForm implements IPageForm {
           }
 
           @Override
-          protected int getConfiguredLabelPosition() {
+          protected byte getConfiguredLabelPosition() {
             return LABEL_POSITION_ON_FIELD;
           }
         }
