+++ code/widgets/org.eclipse.scout.widgets.old.client/src/main/java/org/eclipse/scout/widgets/old/client/ui/forms/Menu1Form.java
@@ -131,7 +131,7 @@ public class Menu1Form extends AbstractForm {
         }
 
         @Override
-        protected int getConfiguredLabelPosition() {
+        protected byte getConfiguredLabelPosition() {
           return LABEL_POSITION_ON_FIELD;
         }
       }
